+++ src/module/lights-hud-ate-models.ts
@@ -108,6 +108,7 @@ export class LightDataHud {
   _id: string;
   flags: any;
   isFlag: boolean;
+  isActorEffect:boolean;
 }
 
 export class LightDataDialog {